Refactor caching strategy of follower counts
During the evolution of follower counts, we had implemented a caching strategy that used Dalli in order to store follower counts so that we wouldn't query the service objects each time we needed to know how many followers a particular had. This commit proposes a new solution to caching follower counts, using columns in the database, which will allow us to sum the follower counts in a callback.

"Media" such as an Instagram or Vine feed are still using Dalli.